Update to the rock pikmin command
Figured out how to use bot embeds. Only thing I couldn't figure out is how to put the color since it asks for it in hex but expects an integer.

I'd also like the send command to send the file as an embed but it's a bit hard because of filenames stuff